+++ src/include/smash/outputformatter.h
@@ -12,6 +12,7 @@
 
 #include <functional>
 #include <map>
+#include <sstream>
 #include <string>
 #include <vector>
 
@@ -19,82 +20,63 @@
 
 namespace smash {
 
-/// Structure to convert a given value into ASCII
-struct ASCII {
+/**
+ * Structure to convert a given value into ASCII format, such that all methods
+ * return a \c std::string.
+ */
+struct ToASCII {
   /// Return type of this converter.
-  using ReturnType = std::string;
+  using type = std::string;
 
   /**
-   * Makes the struct callable. Converts a value into a string with the default
-   * format specifier for the value type.
+   * Converts an integer.
    *
-   * \param[in] value Value to be written.
-   * \return Formatted string.
+   * \param[in] value number to convert
    */
-  template <typename T>
-  std::string operator()(const T& value) const {
-    return std::to_string(value);
+  type as_integer(const int value) const { return std::to_string(value); }
+
+  /**
+   * Converts a double with 6 digits of precision.
+   *
+   * \param[in] value number to convert
+   */
+  type as_double(const double value) const {
+    std::ostringstream os{};
+    os << std::setprecision(6) << value;
+    return os.str();
   }
-  /**
-   * Overload the struct call for when a string is given.
-   *
-   * \param[in] value string to be written.
-   * \return the same string.
-   */
-  std::string operator()(const std::string& value) const { return value; }
-  /**
-   * Overload the struct call for when a literal string (sequence of chars) is
-   * given.
-   *
-   * \param[in] value string to be written.
-   * \return the same string.
-   */
-  std::string operator()(const char* value) const { return value; }
 
   /**
-   * Converts a value into a string with a format specifier, occupying a maximum
-   * of 20 chars.
+   * Converts a double with 9 digits of precision.
    *
-   * \param[in] value Value to be written.
-   * \param[in] format Format specifier to use.
-   * \return Formatted string.
+   * \param[in] value number to convert
    */
-  template <typename T>
-  std::string operator()(const T& value, const char* format) const {
-    // Maximum size of a string is 20.
-    constexpr std::size_t kMaxSize = 20;
-    char buffer[kMaxSize];
-    std::snprintf(buffer, kMaxSize, format, value);
-    return buffer;
+  type as_precise_double(const double value) const {
+    std::ostringstream os{};
+    os << std::setprecision(9) << value;
+    return os.str();
   }
-};
 
-/// Structure to convert a given value into Binary. This is not yet implemented,
-/// so the default constructor is deleted.
-struct Binary {
-  /// Return type of this converter.
-  using ReturnType = const void*;
-
-  // The binary conversion is not yet implemented
-  Binary() = delete;
   /**
-   * Gives the pointer to the stored data with no format specified.
+   * Because %ToASCII converts into strings, this simply returns the string
+   * itself.
    *
-   * \param[in] value Value to be written.
-   * \return a pointer to the value.
+   * \param[inout] str string to be written
    */
-  template <typename T>
-  const void* operator()(const T& value) const {
-    return static_cast<const void*>(&value);
-  }
+  type as_string(const std::string& str) const { return str; }
 };
 
 /**
  * \tparam Converter format desired for the output.
  * A general formatter used for output purposes, which currently only works for
- * ASCII-based formats.
+ * ASCII-based formats. In the future, a Binary converter will be implemented.
+ *
+ * In case further quantities need to be made outputable, one needs to add them
+ * in the constructor with the appropriate getter from ParticleData, as well as
+ * insert the proper pair in the units map.
  */
-template <typename Converter>
+template <typename Converter,
+          std::enable_if_t<std::is_same_v<Converter, ToASCII>, bool> = true>
 class OutputFormatter {
  public:
   /**
@@ -111,101 +93,104 @@ class OutputFormatter {
     for (const std::string& quantity : quantities_) {
       if (quantity == "t") {
         getters_.push_back([this](const ParticleData& in) {
-          return this->converter_(in.position()[0], "%g");
+          return this->converter_.as_double(in.position()[0]);
         });
       } else if (quantity == "x") {
         getters_.push_back([this](const ParticleData& in) {
-          return this->converter_(in.position()[1], "%g");
+          return this->converter_.as_double(in.position()[1]);
         });
       } else if (quantity == "y") {
         getters_.push_back([this](const ParticleData& in) {
-          return this->converter_(in.position()[2], "%g");
+          return this->converter_.as_double(in.position()[2]);
         });
       } else if (quantity == "z") {
         getters_.push_back([this](const ParticleData& in) {
-          return this->converter_(in.position()[3], "%g");
+          return this->converter_.as_double(in.position()[3]);
         });
       } else if (quantity == "mass") {
         getters_.push_back([this](const ParticleData& in) {
-          return this->converter_(in.effective_mass(), "%g");
+          return this->converter_.as_double(in.effective_mass());
         });
       } else if (quantity == "p0") {
         getters_.push_back([this](const ParticleData& in) {
-          return this->converter_(in.momentum()[0], "%.9g");
+          return this->converter_.as_precise_double(in.momentum()[0]);
         });
       } else if (quantity == "px") {
         getters_.push_back([this](const ParticleData& in) {
-          return this->converter_(in.momentum()[1], "%.9g");
+          return this->converter_.as_precise_double(in.momentum()[1]);
         });
       } else if (quantity == "py") {
         getters_.push_back([this](const ParticleData& in) {
-          return this->converter_(in.momentum()[2], "%.9g");
+          return this->converter_.as_precise_double(in.momentum()[2]);
         });
       } else if (quantity == "pz") {
         getters_.push_back([this](const ParticleData& in) {
-          return this->converter_(in.momentum()[3], "%.9g");
+          return this->converter_.as_precise_double(in.momentum()[3]);
         });
       } else if (quantity == "pdg") {
         getters_.push_back([this](const ParticleData& in) {
-          return this->converter_(in.pdgcode().string().c_str(), "%s");
+          return this->converter_.as_string(in.pdgcode().string().c_str());
         });
       } else if (quantity == "ID" || quantity == "id") {
         getters_.push_back([this](const ParticleData& in) {
-          return this->converter_(in.id(), "%i");
+          return this->converter_.as_integer(in.id());
         });
       } else if (quantity == "charge") {
         getters_.push_back([this](const ParticleData& in) {
-          return this->converter_(in.type().charge(), "%i");
+          return this->converter_.as_integer(in.type().charge());
         });
       } else if (quantity == "ncoll") {
         getters_.push_back([this](const ParticleData& in) {
-          return this->converter_(in.get_history().collisions_per_particle,
-                                  "%i");
+          return this->converter_.as_integer(
+              in.get_history().collisions_per_particle);
         });
       } else if (quantity == "form_time") {
         getters_.push_back([this](const ParticleData& in) {
-          return this->converter_(in.formation_time(), "%g");
+          return this->converter_.as_double(in.formation_time());
         });
       } else if (quantity == "xsecfac") {
         getters_.push_back([this](const ParticleData& in) {
-          return this->converter_(in.xsec_scaling_factor(), "%g");
+          return this->converter_.as_double(in.xsec_scaling_factor());
         });
       } else if (quantity == "proc_id_origin") {
         getters_.push_back([this](const ParticleData& in) {
-          return this->converter_(in.get_history().id_process, "%i");
+          return this->converter_.as_integer(in.get_history().id_process);
         });
       } else if (quantity == "proc_type_origin") {
         getters_.push_back([this](const ParticleData& in) {
-          return this->converter_(
-              static_cast<int>(in.get_history().process_type), "%i");
+          return this->converter_.as_integer(
+              static_cast<int>(in.get_history().process_type));
         });
       } else if (quantity == "time_last_coll") {
         getters_.push_back([this](const ParticleData& in) {
-          return this->converter_(in.get_history().time_last_collision, "%g");
+          return this->converter_.as_double(
+              in.get_history().time_last_collision);
         });
       } else if (quantity == "pdg_mother1") {
         getters_.push_back([this](const ParticleData& in) {
-          return this->converter_(in.get_history().p1.string().c_str(), "%s");
+          return this->converter_.as_string(
+              in.get_history().p1.string().c_str());
         });
       } else if (quantity == "pdg_mother2") {
         getters_.push_back([this](const ParticleData& in) {
-          return this->converter_(in.get_history().p2.string().c_str(), "%s");
+          return this->converter_.as_string(
+              in.get_history().p2.string().c_str());
         });
       } else if (quantity == "baryon_number") {
         getters_.push_back([this](const ParticleData& in) {
-          return this->converter_(in.pdgcode().baryon_number(), "%i");
+          return this->converter_.as_integer(in.pdgcode().baryon_number());
         });
       } else if (quantity == "strangeness") {
         getters_.push_back([this](const ParticleData& in) {
-          return this->converter_(in.pdgcode().strangeness(), "%i");
+          return this->converter_.as_integer(in.pdgcode().strangeness());
         });
       } else if (quantity == "spin_projection") {
         getters_.push_back([this](const ParticleData& in) {
-          return this->converter_(in.spin_projection(), "%i");
+          return this->converter_.as_integer(in.spin_projection());
         });
       } else if (quantity == "0") {  // for OSCAR1999
         getters_.push_back([this]([[maybe_unused]] const ParticleData& in) {
-          return this->converter_(0, "%i");
+          return this->converter_.as_integer(0);
         });
       }
     }
@@ -217,59 +202,54 @@ class OutputFormatter {
    * \param[in] p particle whose information is to be written.
    * \return string with formatted data separated by a space.
    */
-  const std::string data_line(const ParticleData& p) {
-    std::string line;
-    for (const auto& getter : getters_) {
-      line += getter(p) + " ";
-    }
-    if (!line.empty()) {
-      line.pop_back();
-      line += "\n";
-    }
-    return line;
+  typename Converter::type data_line(const ParticleData& p) const {
+    return std::accumulate(
+        std::begin(getters_), std::end(getters_), std::string{},
+        [&](const std::string& ss,
+            const std::function<typename Converter::type(const ParticleData&)>
+                getter) {
+          return ss.empty() ? getter(p) : ss + " " + getter(p);
+        });
   }
 
   /**
    * Produces the line with quantities for the header of the output file.
    * \return string with name of quantities separated by a space.
    */
-  const std::string quantities_line() {
-    std::string header;
-    for (const std::string& q : quantities_) {
-      header += this->converter_(q) + " ";
-    }
-    header.pop_back();
-    return header;
+  typename Converter::type quantities_line() const {
+    return std::accumulate(
+        std::begin(quantities_), std::end(quantities_), std::string{},
+        [&](const std::string& ss, const std::string& s) {
+          return ss.empty() ? converter_.as_string(s)
+                            : ss + " " + converter_.as_string(s);
+        });
   }
 
   /**
    * Produces the line with units for the header of the output file.
    * \return string with units separated by a space.
    */
-  std::string unit_line() {
-    std::string line;
-    for (const std::string& q : quantities_) {
-      line += this->converter_(units_.at(q)) + " ";
-    }
-    line.pop_back();
-    return line;
+  typename Converter::type unit_line() const {
+    return std::accumulate(
+        std::begin(quantities_), std::end(quantities_), std::string{},
+        [&](const std::string& ss, const std::string& s) {
+          return ss.empty() ? converter_.as_string(units_.at(s))
+                            : ss + " " + converter_.as_string(units_.at(s));
+        });
   }
 
  private:
   /// Desired format for data conversion. Currently only ASCII is available.
-  Converter converter_;
+  Converter converter_{};
 
   /// List of quantities to be written.
-  std::vector<std::string> quantities_;
+  std::vector<std::string> quantities_{};
 
   /// List of getters for the corresponding quantities.
-  std::vector<
-      std::function<typename Converter::ReturnType(const ParticleData&)>>
-      getters_;
+  std::vector<std::function<typename Converter::type(const ParticleData&)>>
+      getters_{};
 
-  /**
-   *  Map with known quantities and corresponding units.
-   */
+  ///  Map with known quantities and corresponding units.
   const std::map<std::string, std::string> units_ = {
       {"t", "fm"},
       {"x", "fm"},
@@ -299,13 +279,13 @@ class OutputFormatter {
 
   /// Checks whether the quantities requested are known and unique
   void validate_quantities() {
-    std::string error_message;
     if (quantities_.empty()) {
       throw std::invalid_argument(
           "OutputFormatter: Quantities not given, "
           "please fix the configuration file.");
     }
-    std::string repeated;
+    std::string error_message{};
+    std::string repeated{};
     for (const std::string& quantity : quantities_) {
       if (std::count(quantities_.begin(), quantities_.end(), quantity) > 1) {
         repeated += "'" + quantity + "',";
@@ -315,7 +295,7 @@ class OutputFormatter {
       error_message += "OutputFormatter: Repeated quantities: " + repeated +
                        " please fix the configuration file.\n";
     }
-    std::string unknown;
+    std::string unknown{};
     for (const std::string& quantity : quantities_) {
       if (units_.count(quantity) == 0) {
         unknown += "'" + quantity + "',";